+++ src/app/(route)/(auth)/login/page.tsx
@@ -37,9 +37,9 @@ const Page = () => {
 
       {/* divider 구분선 */}
       <div className="flex h-[18px] w-full items-center px-5">
-        <div className="h-px flex-1 bg-[#E4E4E4]" />
+        <hr className="h-px flex-1 bg-[#E4E4E4]" />
         <span className="px-3 text-[12px] text-[#9D9D9D]">또는</span>
-        <div className="h-px flex-1 bg-[#E4E4E4]"></div>
+        <hr className="h-px flex-1 bg-[#E4E4E4]" />
       </div>
 
       {/* 회원확인 여부 */}
